Assistant/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.9
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pypiwin32==223
python-dotenv==1.1.1
pyttsx3==2.99
pywin32==310
requests==2.32.4
rfc3986==1.5.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
